+++ frontend/src/components/modules/characters/modals/panels/characteristics/sections/Background.tsx
